// src/app/model/employee.ts
export class Employee {
    employeeID: number=0;
    employeeName: string='';
    employeeType: string = '';
    gender: string='';
    role: string='';
    skills: string='';
    division: string='';

}
export class CreateEmployee {
    employeeName: string='';
    employeeType: string = '';
    gender: string='';
    role: string='';
    skills: string='';
    division: string='';
}